btrfs-progs: handle split-brain case for scanned changing device with INCOMPAT_METADATA_UUID
For a scanned changing device with INCOMPAT_METADATA_UUID, which means
its fsid and metadata_uuid differs, then situation can be

a) The final changed devices are still own the INCOMPAT_METADATA_UUID
feature. So their metadata_uuid is same as the device's, but fsids must
differ.

b) The new fsid to be changing is same as the scanned device's
metadata_uuid. So the successful synced devices are like some never
changed. fs_devices's fsid is same with their metadata_uuid in memory.
In disk, those synced device's fsid differ from their metadata_uuid,
because the later member is full zero bytes.

c) The device is the newest device, there are old devices with
INCOMPAT_METADATA_UUID or scanned changing devices.

Signed-off-by: Su Yue <Damenly_Su@gmx.com>